pyproject.toml
[tool.poetry]
name = "ambiplayer"
version = "0.1.0"
description = ""
authors = ["Marc Green <marc1701@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
scipy = "^1.11.2"
sounddevice = "^0.4.6"
pysoundfile = "^0.9.0.post1"
numpy = "^1.25.2"
pyqt6 = "^6.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
